+++ library/src/main/java/com/github/jorgecastilloprz/progressarc/ProgressArcDrawable.java
@@ -218,6 +218,7 @@ final class ProgressArcDrawable extends Drawable implements Animatable {
   public void reset() {
     stop();
     resetProperties();
+    setupAnimations();
     start();
   }